+++ src/main/java/net/geforcemods/securitycraft/blocks/SecurityCameraBlock.java
@@ -112,7 +112,7 @@ public class SecurityCameraBlock extends DisguisableBlock implements IWaterLogga
 
 	@Override
 	public VoxelShape getShape(BlockState state, IBlockReader level, BlockPos pos, ISelectionContext ctx) {
-		BlockState disguisedState = IDisguisable.getDisguisedStateOrDefault(state, level, pos);
+		BlockState disguisedState = IDisguisable.getDisguisedBlockState(level.getBlockEntity(pos)).orElse(state);
 
 		if (disguisedState.getBlock() != this)
 			return disguisedState.getShape(level, pos, ctx);