+++ Changelog.md
@@ -1,6 +1,7 @@
 --------------------------Changelog for the upcoming version of SecurityCraft--------------------------
 
 - Change: The block mine overlay now shows in the first person hand models as well
+- API: IDisguisable has been simplified
 - Fix: Crash when triggering the passcode spam log warning on a server
 - Fix: Reinforced Mud Bricks and Blackstone Slab/Stairs don't look 100% like their vanilla counterparts